# `.data` pronoun speedup